# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TagPulse",
    "description": "IoT platform for RFID tag readers and sensor data",
    "version": "0.1.0"
  },
  "paths": {
...
    "/telemetry": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Ingest Telemetry",
        "description": "Ingest a batch of telemetry readings via HTTP push.",
        "operationId": "ingest_telemetry_telemetry_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Ingest Telemetry Telemetry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Telemetry",
        "description": "Query persisted telemetry readings with filters.",
        "operationId": "list_telemetry_telemetry_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "name": "metric_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric Name"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetryResponse"
                  },
                  "title": "Response List Telemetry Telemetry Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/telemetry-models": {
      "get": {
        "tags": [
          "telemetry-models"
        ],
        "summary": "List Telemetry Models",
        "description": "List all telemetry model definitions.",
        "operationId": "list_telemetry_models_telemetry_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TelemetryModelResponse"
                  },
                  "type": "array",
                  "title": "Response List Telemetry Models Telemetry Models Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "telemetry-models"
        ],
        "summary": "Create Telemetry Model",
        "description": "Define the telemetry schema for a device type.",
        "operationId": "create_telemetry_model_telemetry_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryModelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/telemetry-models/{device_type}": {
      "get": {
        "tags": [
          "telemetry-models"
        ],
        "summary": "Get Telemetry Model",
        "description": "Get telemetry model definition for a device type.",
        "operationId": "get_telemetry_model_telemetry_models__device_type__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "device_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/telemetry-models/{model_id}": {
      "delete": {
        "tags": [
          "telemetry-models"
        ],
        "summary": "Delete Telemetry Model",
        "description": "Delete a telemetry model definition.",
        "operationId": "delete_telemetry_model_telemetry_models__model_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetricDefinition": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "unit": {
            "type": "string",
            "maxLength": 50,
            "title": "Unit"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "unit"
        ],
        "title": "MetricDefinition",
        "description": "A single metric that a device type can report."
      },
...
      "TelemetryBatch": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/TelemetryReading"
            },
            "type": "array",
            "minItems": 1,
            "title": "Readings"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "readings"
        ],
        "title": "TelemetryBatch",
        "description": "Batched telemetry payload \u2014 HTTP and MQTT share this shape."
      },
      "TelemetryModelCreate": {
        "properties": {
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricDefinition"
            },
            "type": "array",
            "minItems": 1,
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "device_type",
          "metrics"
        ],
        "title": "TelemetryModelCreate",
        "description": "Define the telemetry schema for a device type."
      },
      "TelemetryModelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricDefinition"
            },
            "type": "array",
            "title": "Metrics"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_type",
          "metrics",
          "created_at",
          "updated_at"
        ],
        "title": "TelemetryModelResponse",
        "description": "Telemetry model definition returned from the API."
      },
...
      "TelemetryReading": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "metric_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Metric Name"
          },
          "metric_value": {
            "type": "number",
            "title": "Metric Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "metric_name",
          "metric_value"
        ],
        "title": "TelemetryReading",
        "description": "A single telemetry reading inside a batched payload."
      },
      "TelemetryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "metric_name": {
            "type": "string",
            "title": "Metric Name"
          },
          "metric_value": {
            "type": "number",
            "title": "Metric Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "timestamp",
          "metric_name",
          "metric_value",
          "unit"
        ],
        "title": "TelemetryResponse",
        "description": "Persisted telemetry row."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID"
      }
    }
  }
}